Refactor ModuleIdentifier out into its own class
